requirements.txt
absl-py==1.4.0
accelerate==0.22.0
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.5.0
antlr4-python3-runtime==4.8
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chumpy==0.70
click==8.1.7
cmake==3.27.2
comm==0.1.4
ConfigArgParse==1.7
configer==1.3.1
configparser==6.0.0
contourpy==1.1.0
controlnet-aux==0.0.6
croniter==1.3.15
cycler==0.11.0
Cython==3.0.2
dateutils==0.6.12
debugpy==1.6.7.post1
decorator==5.1.1
deepdiff==6.3.1
defusedxml==0.7.1
deprecation==2.1.0
diff-gaussian-rasterization==0.0.0
docker-pycreds==0.4.0
einops==0.6.1
entrypoints==0.4
envlight==0.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.88.0
fastjsonschema==2.18.0
filelock==3.12.3
flask==2.3.3
fonttools==4.42.1
freetype-py==2.4.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
h11==0.14.0
h5py==3.9.0
huggingface-hub==0.16.4
human-body-prior==0.8.5.0
hydra-core==1.1.2
idna==3.4
imageio==2.28.0
importlib-metadata==6.8.0
importlib-resources==5.2.3
inquirer==3.1.3
iopath==0.1.10
ipycanvas==0.13.1
ipyevents==2.0.2
ipykernel==6.25.1
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.0
itsdangerous==2.1.2
jaxtyping==0.2.19
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-client==7.4.9
jupyter-core==5.3.1
jupyter-events==0.7.0
jupyter-packaging==0.12.3
jupyter-server==2.7.3
jupyter-server-fileid==0.9.0
jupyter-server-terminals==0.4.4
jupyter-server-ydoc==0.8.0
jupyter-ydoc==0.2.5
jupyterlab==3.6.5
jupyterlab-pygments==0.2.2
jupyterlab-server==2.24.0
jupyterlab-widgets==3.0.8
kaolin==0.14.0
kiwisolver==1.4.5
kornia==0.7.0
lap==0.4.0
lazy-loader==0.3
libigl==2.4.1
lightning==2.0.0
lightning-cloud==0.5.37
lightning-utilities==0.9.0
lit==16.0.6
lpips==0.1.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nerfacc==0.5.3
nest-asyncio==1.5.7
networkx==3.1
ninja==1.11.1
notebook==6.5.5
notebook-shim==0.2.3
numpy==1.23.1
nvdiffrast==0.3.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.1.2
open3d==0.13.0
opencv-python==4.8.0.76
ordered-set==4.1.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
plotly==5.18.0
point-cloud-utils==0.30.0
portalocker==2.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
pydantic==2.3.0
pydantic-core==2.6.3
pyDeprecate==0.3.1
pyglet==2.0.9
Pygments==2.16.1
PyJWT==2.8.0
PyMCubes==0.1.4
pymeshlab==2022.2.post3
PyOpenGL==3.1.5
pyparsing==3.0.9
pyrender==0.1.45
pysdf==0.1.9
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==1.5.7
pytorch3d==0.7.4
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==24.0.1
readchar==4.0.5
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.10.2
rsa==4.9
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.0
scikit-video==1.1.11
scipy==1.10.1
segment-anything==1.0
Send2Trash==1.8.2
sentry-sdk==1.30.0
setproctitle==1.3.2
simple-romp==1.1.3
six==1.16.0
smmap==5.0.0
smplx==0.1.28
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.31.1
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.6
tinycss2==1.2.1
tinycudann==1.6
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.1
torch==2.0.1
torch-tb-profiler==0.4.1
torchgeometry==0.1.2
torchmetrics==1.1.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.32.1
transforms3d==0.4.1
trimesh==3.23.5
triton==2.0.0
typeguard==4.1.3
typing-extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
usd-core==23.5
uvicorn==0.23.2
wandb==0.15.9
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
werkzeug==2.3.7
widgetsnbextension==4.0.8
xformers==0.0.21
y-py==0.6.0
yacs==0.1.8
yarl==1.9.2
ypy-websocket==0.8.4
zipp==3.16.2
